Unit Tests
This commit introduces a small test suite designed to exercise the
functionality of the the Vault Service Broker in a variety of real-world
scenarios.  It does so by standing up a development Vault server and a
test instance of the vault-broker process, and interacting with both via
tools like safe and curl.

This is a framework and a starting configuration of tests.  Expect more
as bugs are discovered.

Fixes #1