High Availability Support (#10)
IPs and Advertise URLs are provided in a similar fashion to the current implementation.
If they are set and the broker is unable to contact the vault via the first address in the list (or the `VAULT_ADDR` that was provided) it will failover to the next vault in the list. This allows operators to set precedence (if applicable).

Once a vault is marked as failed, future requests are diverted to alternate vaults until the health check routine finds that they are back online.